// src/app/api/shared/layout/count-response.interface.ts
export interface CountResponseInterface {
  lead: number;
  lead_to_answer: number;
  mailbox: number;
  matching_contact: number;
  matching_property: number;
  matching_promotion: number;
  contact_to_follow_up: number;
  contact_invalid_email: number;
  contact_no_search: number;
  report_validation: number;
}